Fixed GetCpuTime crash. Changed IStructure.Hits and IStructure.HitsMax to be nullable as some arena structures don't have hits. Fixed pathfinder paths copied to wasm incorrectly. Fixed missing visual bindings.

+++ ScreepsDotNet.API/API/Arena/IUtils.cs
@@ -175,7 +175,7 @@ namespace ScreepsDotNet.API.Arena
         /// Get CPU wall time elapsed in the current tick in nanoseconds
         /// </summary>
         /// <returns></returns>
-        long GetCpuTime();
+        double GetCpuTime();
 
         /// <summary>
         /// Get linear direction by differences of x and y.
